updated 08/18/2022
returns two messages in different cases. if this wallpaper has already been installed, then the program warns the user about this (hex format), if there were no such wallpapers yet, then it warns (also hex format) about this and installs them

+++ main.py
@@ -1,5 +1,6 @@
 import requests
 from bs4 import BeautifulSoup as bs
+import os.path
 
 
 def getImageFile():
@@ -11,11 +12,17 @@ def getImageFile():
 	image_url = temp[temp.find('/'):temp.find(';')-1]
 
 	temp = image_url[image_url.find('id=')+3:image_url.find('.jpg')]
-	short_name = temp[temp.find('.')+1:temp.find('_')]
+	image_name = temp[temp.find('.')+1:temp.find('_')]
+	if os.path.isfile(f'wallpapers/{image_name}.jpg'):
+		return '7468652077616c6c70617065722068617320616c7265616479206265656e20646f776e6c6f61646564'
 
 	image_data = requests.get(URL+image_url).content
 	with open(f'wallpapers/{image_name}.jpg', 'wb') as file:
 		file.write(image_data)
 
+	return '6e65772077616c6c7061706572732068617665206265656e20696e7374616c6c6564'
 
-getImageFile()
+
+
+
+print(getImageFile())
